refactor: Move frontend factoring into its own file and use in Random
  Resolves #154.
